Day 18. Puzzle 2.
DONE!

The input data had zero-based co-ords, which I hadn't accounted for -
thinking that I'd checked all input was 1-based, sigh.

Most of the code in this commit is not needed now, but repreesents what
I had to add to find the bug :p line 182 contains the only important
change.